requirements_benchmark.txt
# Основные зависимости для PolitAgent Benchmark System

# Анализ данных и визуализация
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.21.0

# Конфигурация
pyyaml>=6.0

# Асинхронность и параллельность
asyncio-throttle>=1.0.2

# Логирование и утилиты
tqdm>=4.64.0

# Опциональные зависимости для расширенной функциональности
# Раскомментируйте при необходимости:

# plotly>=5.0.0          # Интерактивные графики
# jupyter>=1.0.0         # Jupyter notebooks для анализа
# scikit-learn>=1.1.0    # ML метрики и анализ
# scipy>=1.9.0           # Статистический анализ 